scripts/requirements.txt
opencv-python==4.10.0.84
numpy==2.1.1
Pillow==10.4.0
